edge_encoder/requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
numpy
Pillow
